Major change to command names to instead use migrateprevious and migr… (#2)
Major change to command names to instead use migrateprevious and migraterollback to follow the more traditional Django command format. Also opted for positional arguments for the same reason. Also used the internal management to call migrate in order to correctly show the command being run. Updated README.md to reflect this and setup.py to have correct values when publishing to PyPi.